% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spatialDmelxsim}
\alias{spatialDmelxsim}
\title{Spatial allelic expression counts for fly cross embryo}
\description{
Allelic expression counts of spatial slices of a fly embryo from
Combs & Fraser (2018), a D melanogaster x D simulans reciprocal cross
}
\examples{

suppressPackageStartupMessages(library(SummarizedExperiment))
se <- spatialDmelxsim()
se

}
\references{
Combs PA, Fraser HB (2018) Spatially varying cis-regulatory divergence
in Drosophila embryos elucidates cis-regulatory logic.
PLOS Genetics 14(11): e1007631.
}
